Spec 0013 — Move Base/Casts/* → Casts/
First category of the Base/ directory drain. Relocates the five cast
classes to a top-level Casts/ folder and rewrites every use site across
the monorepo. Updates the v1→v2 Rector renames so downstream consumers
follow.
