+++ packages/core/src/Casts/ShippingBreakdown.php
@@ -1,16 +1,16 @@
 <?php
 
-namespace Lunar\Core\Base\Casts;
+namespace Lunar\Core\Casts;
 
 use Illuminate\Contracts\Database\Eloquent\CastsAttributes;
 use Illuminate\Contracts\Database\Eloquent\SerializesCastableAttributes;
 use Illuminate\Database\Eloquent\Model;
-use Lunar\Core\Base\ValueObjects\Cart\TaxBreakdownAmount;
+use Illuminate\Support\Collection;
+use Lunar\Core\Base\ValueObjects\Cart\ShippingBreakdownItem;
 use Lunar\Core\DataObjects\PriceValue;
 use Lunar\Core\Models\Currency;
-use Spatie\LaravelBlink\BlinkFacade;
 
-class TaxBreakdown implements CastsAttributes, SerializesCastableAttributes
+class ShippingBreakdown implements CastsAttributes, SerializesCastableAttributes
 {
     /**
      * Cast the given value.
@@ -19,25 +19,22 @@ class TaxBreakdown implements CastsAttributes, SerializesCastableAttributes
      * @param  string  $key
      * @param  mixed  $value
      * @param  array  $attributes
-     * @return \Lunar\Core\Base\ValueObjects\Cart\TaxBreakdown
+     * @return \Lunar\Core\Base\ValueObjects\Cart\ShippingBreakdown
      */
     public function get($model, $key, $value, $attributes)
     {
-        $breakdown = new \Lunar\Core\Base\ValueObjects\Cart\TaxBreakdown;
+        $breakdown = new \Lunar\Core\Base\ValueObjects\Cart\ShippingBreakdown;
 
-        $breakdown->amounts = collect(
+        $breakdown->items = collect(
             json_decode($value, false)
-        )->mapWithKeys(function ($amount, $key) {
-            $currency = BlinkFacade::once("currency_{$amount->currency_code}", function () use ($amount) {
-                return Currency::whereCode($amount->currency_code)->first();
-            });
+        )->mapWithKeys(function ($shipping, $key) {
+            $currency = Currency::whereCode($shipping->currency->code)->first();
 
             return [
-                $key => new TaxBreakdownAmount(
-                    price: new PriceValue((int) $amount->value, $currency),
-                    identifier: $amount->identifier,
-                    description: $amount->description,
-                    percentage: $amount->percentage,
+                $key => new ShippingBreakdownItem(
+                    name: $shipping->name,
+                    identifier: $shipping->identifier,
+                    price: new PriceValue((int) $shipping->value, $currency),
                 ),
             ];
         });
@@ -50,16 +47,14 @@ class TaxBreakdown implements CastsAttributes, SerializesCastableAttributes
      *
      * @param  Model  $model
      * @param  string  $key
-     * @param  \Lunar\Core\Base\ValueObjects\Cart\TaxBreakdown  $value
+     * @param  \Lunar\Core\Base\ValueObjects\Cart\ShippingBreakdown  $value
      * @param  array  $attributes
      * @return array
-     *
-     * @throws \Exception
      */
     public function set($model, $key, $value, $attributes)
     {
-        if ($value && ! is_a($value, \Lunar\Core\Base\ValueObjects\Cart\TaxBreakdown::class)) {
-            throw new \Exception('Tax breakdown must be instance of Lunar\Core\Base\ValueObjects\Cart\TaxBreakdown');
+        if ($value && ! is_a($value, \Lunar\Core\Base\ValueObjects\Cart\ShippingBreakdown::class)) {
+            throw new \Exception('Shipping breakdown must be instance of Lunar\Core\Base\ValueObjects\Cart\ShippingBreakdown');
         }
 
         if (! $value) {
@@ -67,13 +62,15 @@ class TaxBreakdown implements CastsAttributes, SerializesCastableAttributes
         }
 
         return [
-            $key => $value->amounts->map(function ($item) {
+            $key => $value->items->map(function ($item) {
+                $currency = $item->price->resolveCurrency();
+
                 return [
-                    'description' => $item->description,
+                    'name' => $item->name,
                     'identifier' => $item->identifier,
-                    'percentage' => $item->percentage,
                     'value' => $item->price->value,
-                    'currency_code' => $item->price->resolveCurrency()->code,
+                    'formatted' => $item->price->format(),
+                    'currency' => $currency->toArray(),
                 ];
             })->toJson(),
         ];
@@ -84,7 +81,7 @@ class TaxBreakdown implements CastsAttributes, SerializesCastableAttributes
      *
      * @param  Model  $model
      * @param  string  $key
-     * @param  mixed  $value
+     * @param  Collection  $value
      * @param  array<string, mixed>  $attributes
      */
     public function serialize($model, $key, $value, $attributes)
