+++ packages/core/src/Models/Brand.php
@@ -7,7 +7,6 @@ use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 use Illuminate\Database\Eloquent\Relations\HasMany;
 use Illuminate\Support\Carbon;
 use Lunar\Core\Base\BaseModel;
-use Lunar\Core\Base\Casts\AsAttributeData;
 use Lunar\Core\Base\Traits\HasAttributes;
 use Lunar\Core\Base\Traits\HasMacros;
 use Lunar\Core\Base\Traits\HasMedia;
@@ -15,6 +14,7 @@ use Lunar\Core\Base\Traits\HasTranslations;
 use Lunar\Core\Base\Traits\HasUrls;
 use Lunar\Core\Base\Traits\LogsActivity;
 use Lunar\Core\Base\Traits\Searchable;
+use Lunar\Core\Casts\AsAttributeData;
 use Lunar\Core\Database\Factories\BrandFactory;
 use Lunar\Core\Facades\DB;
 use Spatie\MediaLibrary\HasMedia as SpatieHasMedia;
