+++ packages/core/src/Models/Cart.php
@@ -24,7 +24,6 @@ use Lunar\Core\Actions\Carts\SetShippingOption;
 use Lunar\Core\Actions\Carts\UpdateCartLine;
 use Lunar\Core\Base\Addressable;
 use Lunar\Core\Base\BaseModel;
-use Lunar\Core\Base\Casts\CouponString;
 use Lunar\Core\Base\LunarUser;
 use Lunar\Core\Base\Purchasable;
 use Lunar\Core\Base\Traits\CachesProperties;
@@ -35,6 +34,7 @@ use Lunar\Core\Base\ValueObjects\Cart\FreeItem;
 use Lunar\Core\Base\ValueObjects\Cart\Promotion;
 use Lunar\Core\Base\ValueObjects\Cart\ShippingBreakdown;
 use Lunar\Core\Base\ValueObjects\Cart\TaxBreakdown;
+use Lunar\Core\Casts\CouponString;
 use Lunar\Core\Database\Factories\CartFactory;
 use Lunar\Core\DataObjects\PriceValue;
 use Lunar\Core\DataTypes\ShippingOption;
