+++ packages/core/src/Models/Collection.php
@@ -10,7 +10,6 @@ use Illuminate\Support\Carbon;
 use Kalnoy\Nestedset\NodeTrait;
 use Kalnoy\Nestedset\QueryBuilder;
 use Lunar\Core\Base\BaseModel;
-use Lunar\Core\Base\Casts\AsAttributeData;
 use Lunar\Core\Base\HasThumbnailImage;
 use Lunar\Core\Base\Traits\HasChannels;
 use Lunar\Core\Base\Traits\HasCustomerGroups;
@@ -19,6 +18,7 @@ use Lunar\Core\Base\Traits\HasMedia;
 use Lunar\Core\Base\Traits\HasTranslations;
 use Lunar\Core\Base\Traits\HasUrls;
 use Lunar\Core\Base\Traits\Searchable;
+use Lunar\Core\Casts\AsAttributeData;
 use Lunar\Core\Database\Factories\CollectionFactory;
 use Spatie\MediaLibrary\HasMedia as SpatieHasMedia;
