+++ packages/core/src/Models/Customer.php
@@ -9,13 +9,13 @@ use Illuminate\Database\Eloquent\Relations\HasMany;
 use Illuminate\Database\Eloquent\Relations\MorphToMany;
 use Illuminate\Support\Carbon;
 use Lunar\Core\Base\BaseModel;
-use Lunar\Core\Base\Casts\AsAttributeData;
 use Lunar\Core\Base\Traits\HasAttributes;
 use Lunar\Core\Base\Traits\HasMacros;
 use Lunar\Core\Base\Traits\HasPersonalDetails;
 use Lunar\Core\Base\Traits\HasTranslations;
 use Lunar\Core\Base\Traits\LogsActivity;
 use Lunar\Core\Base\Traits\Searchable;
+use Lunar\Core\Casts\AsAttributeData;
 use Lunar\Core\Database\Factories\CustomerFactory;
 
 /**
