+++ packages/core/src/Models/Order.php
@@ -9,13 +9,13 @@ use Illuminate\Database\Eloquent\Relations\HasMany;
 use Illuminate\Database\Eloquent\Relations\HasOne;
 use Illuminate\Support\Carbon;
 use Lunar\Core\Base\BaseModel;
-use Lunar\Core\Base\Casts\DiscountBreakdown;
-use Lunar\Core\Base\Casts\ShippingBreakdown;
-use Lunar\Core\Base\Casts\TaxBreakdown;
 use Lunar\Core\Base\Traits\HasMacros;
 use Lunar\Core\Base\Traits\HasTags;
 use Lunar\Core\Base\Traits\LogsActivity;
 use Lunar\Core\Base\Traits\Searchable;
+use Lunar\Core\Casts\DiscountBreakdown;
+use Lunar\Core\Casts\ShippingBreakdown;
+use Lunar\Core\Casts\TaxBreakdown;
 use Lunar\Core\Contracts\HasCurrency;
 use Lunar\Core\Database\Factories\OrderFactory;
 use Lunar\Core\Models\Concerns\FormatsPrices;
