+++ packages/core/src/Models/OrderLine.php
@@ -9,9 +9,9 @@ use Illuminate\Database\Eloquent\Relations\HasOneThrough;
 use Illuminate\Database\Eloquent\Relations\MorphTo;
 use Illuminate\Support\Carbon;
 use Lunar\Core\Base\BaseModel;
-use Lunar\Core\Base\Casts\TaxBreakdown;
 use Lunar\Core\Base\Traits\HasMacros;
 use Lunar\Core\Base\Traits\LogsActivity;
+use Lunar\Core\Casts\TaxBreakdown;
 use Lunar\Core\Contracts\HasCurrency;
 use Lunar\Core\Database\Factories\OrderLineFactory;
 use Lunar\Core\Models\Concerns\FormatsPrices;
