+++ packages/core/src/Models/Product.php
@@ -15,7 +15,6 @@ use Illuminate\Database\Eloquent\Relations\MorphMany;
 use Illuminate\Database\Eloquent\SoftDeletes;
 use Illuminate\Support\Carbon;
 use Lunar\Core\Base\BaseModel;
-use Lunar\Core\Base\Casts\AsAttributeData;
 use Lunar\Core\Base\Enums\Concerns\ProvidesProductAssociationType;
 use Lunar\Core\Base\HasThumbnailImage;
 use Lunar\Core\Base\Traits\HasChannels;
@@ -27,6 +26,7 @@ use Lunar\Core\Base\Traits\HasTranslations;
 use Lunar\Core\Base\Traits\HasUrls;
 use Lunar\Core\Base\Traits\LogsActivity;
 use Lunar\Core\Base\Traits\Searchable;
+use Lunar\Core\Casts\AsAttributeData;
 use Lunar\Core\Database\Factories\ProductFactory;
 use Lunar\Core\Jobs\Products\Associations\Associate;
 use Lunar\Core\Jobs\Products\Associations\Dissociate;
