+++ packages/core/src/Models/ProductVariant.php
@@ -9,7 +9,6 @@ use Illuminate\Database\Eloquent\SoftDeletes;
 use Illuminate\Support\Carbon;
 use Illuminate\Support\Collection;
 use Lunar\Core\Base\BaseModel;
-use Lunar\Core\Base\Casts\AsAttributeData;
 use Lunar\Core\Base\HasThumbnailImage;
 use Lunar\Core\Base\Purchasable;
 use Lunar\Core\Base\Traits\HasAttributes;
@@ -18,6 +17,7 @@ use Lunar\Core\Base\Traits\HasMacros;
 use Lunar\Core\Base\Traits\HasPrices;
 use Lunar\Core\Base\Traits\HasTranslations;
 use Lunar\Core\Base\Traits\LogsActivity;
+use Lunar\Core\Casts\AsAttributeData;
 use Lunar\Core\Database\Factories\ProductVariantFactory;
 use Spatie\LaravelBlink\BlinkFacade as Blink;
 use Spatie\MediaLibrary\MediaCollections\Models\Media;
