+++ packages/upgrade/src/Rector/LunarSetList.php
@@ -52,12 +52,12 @@ final class LunarSetList
         'Lunar\\Base\\CartModifier' => 'Lunar\\Core\\Base\\CartModifier',
         'Lunar\\Base\\CartModifiers' => 'Lunar\\Core\\Base\\CartModifiers',
         'Lunar\\Base\\CartSessionInterface' => 'Lunar\\Core\\Base\\CartSessionInterface',
-        'Lunar\\Base\\Casts\\AsAttributeData' => 'Lunar\\Core\\Base\\Casts\\AsAttributeData',
-        'Lunar\\Base\\Casts\\CouponString' => 'Lunar\\Core\\Base\\Casts\\CouponString',
-        'Lunar\\Base\\Casts\\DiscountBreakdown' => 'Lunar\\Core\\Base\\Casts\\DiscountBreakdown',
+        'Lunar\\Base\\Casts\\AsAttributeData' => 'Lunar\\Core\\Casts\\AsAttributeData',
+        'Lunar\\Base\\Casts\\CouponString' => 'Lunar\\Core\\Casts\\CouponString',
+        'Lunar\\Base\\Casts\\DiscountBreakdown' => 'Lunar\\Core\\Casts\\DiscountBreakdown',
         'Lunar\\Base\\Casts\\Price' => 'Lunar\\Core\\Base\\Casts\\Price',
-        'Lunar\\Base\\Casts\\ShippingBreakdown' => 'Lunar\\Core\\Base\\Casts\\ShippingBreakdown',
-        'Lunar\\Base\\Casts\\TaxBreakdown' => 'Lunar\\Core\\Base\\Casts\\TaxBreakdown',
+        'Lunar\\Base\\Casts\\ShippingBreakdown' => 'Lunar\\Core\\Casts\\ShippingBreakdown',
+        'Lunar\\Base\\Casts\\TaxBreakdown' => 'Lunar\\Core\\Casts\\TaxBreakdown',
         'Lunar\\Base\\DataTransferObjects\\CartDiscount' => 'Lunar\\Core\\Base\\DataTransferObjects\\CartDiscount',
         'Lunar\\Base\\DataTransferObjects\\PaymentAuthorize' => 'Lunar\\Core\\Base\\DataTransferObjects\\PaymentAuthorize',
         'Lunar\\Base\\DataTransferObjects\\PaymentCapture' => 'Lunar\\Core\\Base\\DataTransferObjects\\PaymentCapture',
